+++ src/StorageBase/Action/PortalNodeAlias/Find/PortalNodeAliasFindCriteria.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\Base\Action\PortalNodeAlias\Find;
+
+use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
+
+final class PortalNodeAliasFindCriteria implements AttachmentAwareInterface
+{
+    use AttachmentAwareTrait;
+
+    /**
+     * @var string[]
+     */
+    private array $alias;
+
+    /**
+     * @param string[] $alias
+     */
+    public function __construct(array $alias)
+    {
+        $this->attachments = new AttachmentCollection();
+        $this->alias = $alias;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getAlias(): array
+    {
+        return $this->alias;
+    }
+
+    /**
+     * @param string[] $alias
+     */
+    public function setAlias(array $alias): void
+    {
+        $this->alias = $alias;
+    }
+}