+++ src/StorageBase/Action/PortalNodeAlias/Set/PortalNodeAliasSetPayload.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\Base\Action\PortalNodeAlias\Set;
+
+use Heptacom\HeptaConnect\Dataset\Base\AttachmentCollection;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Create\CreatePayloadInterface;
+
+final class PortalNodeAliasSetPayload implements AttachmentAwareInterface, CreatePayloadInterface
+{
+    use AttachmentAwareTrait;
+
+    private PortalNodeKeyInterface $portalNodeKey;
+
+    private ?string $alias;
+
+    public function __construct(PortalNodeKeyInterface $portalNodeKey, ?string $alias)
+    {
+        $this->attachments = new AttachmentCollection();
+        $this->portalNodeKey = $portalNodeKey;
+        $this->alias = $alias;
+    }
+
+    public function getPortalNodeKey(): PortalNodeKeyInterface
+    {
+        return $this->portalNodeKey;
+    }
+
+    public function setPortalNodeKey(PortalNodeKeyInterface $portalNodeKey): void
+    {
+        $this->portalNodeKey = $portalNodeKey;
+    }
+
+    public function getAlias(): ?string
+    {
+        return $this->alias;
+    }
+
+    public function setAlias(?string $alias): void
+    {
+        $this->alias = $alias;
+    }
+}